refactor: api destination helper
